+++ BlueSheep/Common/Protocol/messages/connection/IdentificationAccountForceMessage.cs
@@ -32,7 +32,7 @@ namespace BlueSheep.Common.Protocol.Messages
         {
         }
         
-        public IdentificationAccountForceMessage(bool autoconnect, bool useCertificate, bool useLoginToken, Types.VersionExtended version, string lang, sbyte[] credentials, short serverId, double sessionOptionalSalt, string forcedAccountLogin, List<short> failedAttempts)
+        public IdentificationAccountForceMessage(bool autoconnect, bool useCertificate, bool useLoginToken, Types.VersionExtended version, string lang, sbyte[] credentials, int serverId, double sessionOptionalSalt, string forcedAccountLogin, List<int> failedAttempts)
          : base(autoconnect, useCertificate, useLoginToken, version, lang, credentials, serverId, sessionOptionalSalt, failedAttempts)
         {
             this.forcedAccountLogin = forcedAccountLogin;
