+++ BlueSheep/Common/Protocol/messages/game/achievement/AchievementRewardErrorMessage.cs
@@ -26,20 +26,20 @@ namespace BlueSheep.Common.Protocol.Messages
             get { return ID; }
         }
         
-        public short achievementId;
+        public int achievementId;
         
         public AchievementRewardErrorMessage()
         {
         }
         
-        public AchievementRewardErrorMessage(short achievementId)
+        public AchievementRewardErrorMessage(int achievementId)
         {
             this.achievementId = achievementId;
         }
         
         public override void Serialize(BigEndianWriter writer)
         {
-            writer.WriteShort(achievementId);
+            writer.WriteShort((short)achievementId);
         }
         
         public override void Deserialize(BigEndianReader reader)
