+++ BlueSheep/Common/Protocol/messages/game/achievement/AchievementRewardRequestMessage.cs
@@ -26,20 +26,20 @@ namespace BlueSheep.Common.Protocol.Messages
             get { return ID; }
         }
         
-        public short achievementId;
+        public int achievementId;
         
         public AchievementRewardRequestMessage()
         {
         }
         
-        public AchievementRewardRequestMessage(short achievementId)
+        public AchievementRewardRequestMessage(int achievementId)
         {
             this.achievementId = achievementId;
         }
         
         public override void Serialize(BigEndianWriter writer)
         {
-            writer.WriteShort(achievementId);
+            writer.WriteShort((short)achievementId);
         }
         
         public override void Deserialize(BigEndianReader reader)
