+++ BlueSheep/Common/Protocol/messages/game/context/roleplay/MapComplementaryInformationsDataInHouseMessage.cs
@@ -32,7 +32,7 @@ namespace BlueSheep.Common.Protocol.Messages
         {
         }
         
-        public MapComplementaryInformationsDataInHouseMessage(short subAreaId, int mapId, Types.HouseInformations[] houses, Types.GameRolePlayActorInformations[] actors, Types.InteractiveElement[] interactiveElements, Types.StatedElement[] statedElements, Types.MapObstacle[] obstacles, Types.FightCommonInformations[] fights, Types.HouseInformationsInside currentHouse)
+        public MapComplementaryInformationsDataInHouseMessage(int subAreaId, int mapId, Types.HouseInformations[] houses, Types.GameRolePlayActorInformations[] actors, Types.InteractiveElement[] interactiveElements, Types.StatedElement[] statedElements, Types.MapObstacle[] obstacles, Types.FightCommonInformations[] fights, Types.HouseInformationsInside currentHouse)
          : base(subAreaId, mapId, houses, actors, interactiveElements, statedElements, obstacles, fights)
         {
             this.currentHouse = currentHouse;
